+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.6
+
+### Added
+- `Cassie::Support::SystemCommand.exist?(command)` and `Cassie::Support::SystemCommand#exist?`
+- `Cassie::Support::SystemCommand.which(command)` and `Cassie::Support::SystemCommand#which`
+
 ## 1.1.5
 
 ### Fixed